pipe engine eval_batch: add option to disable loss broadcast (#4326)
it is sometimes not required to have the loss bcasted to all rank after
evaluation cycle and it is only required by some ranks.
It adds overhead of communication between rank.
by setting bcast_loss=False (default is True and retains the previous behavior),
loss will not be bcasted.
if monitor is enabled loss will be bcasted.

Co-authored-by: Logan Adams <114770087+loadams@users.noreply.github.com>