Update node16 check on self-hosted runners and remove python 3.6 (#5756)
With changes from GitHub [finally
deprecating](https://github.com/actions/checkout/issues/1474) [node16
based
runners](https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/)
(which the checkout@v3 action uses) we need to make changes to support
this.

To do this, there are two changes. First we remove the python 3.6 check
as with the changes in pydantic v2 that will be merged soon, we will be
removing this check there, so we can more easily remove it now so that
future PRs are cleaner and it is clear why some changes have been made.

Second, node16 is the default on some of our self-hosted runners. To
work around tests failing on these, we [set the GitHub env var to
override this
check](https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/).

Other relevant links:
https://github.com/actions/checkout/issues/1474
https://github.com/easybuilders/easybuild-framework/pull/4574/files
https://github.com/actions/checkout/issues/1809
https://github.com/actions/runner/issues/3373
https://github.com/actions/checkout/issues/1809