Avoid graph breaks caused by inner classes in the backward hooks
This PR is part of the effort to improve Deepspeed
performance when using PyTorch compile.

There is a known bug in torch.compile
(https://github.com/pytorch/pytorch/issues/128942)
which causes a graph break when a inner class
is defined within a method that is being compiled
and logs the following message:

[__graph_breaks] torch._dynamo.exc.Unsupported: missing: LOAD_BUILD_CLASS

This is the case with the inner classes
PreBackwardFunctionForModule and PostBackwardFunctionModule.

While there is an open PR for this
(https://github.com/pytorch/pytorch/pull/133805),
we can solve the issue by moving the inner classes
into the initialization code.

No graph breaks and the corresponding logs
are produced anymore.

Signed-off-by: Max Kovalenko <mkovalenko@habana.ai>