Added device detection to communication logging
In `comms_logging.py`, when calling log_all and the `show_straggler` option is enabled, an all_reduce is performed across all nodes to calculate the minimum latency to find stragglers.
However, the tensors on which this is performed are not sent to the configured devices. This commit adds this capability using deepspeed's abstract accelerator api.

Signed-off-by: Alex Kiefer <alexkiefer51@gmail.com>