Enhancements to the communication logger and api
The current implementation of the communication logging api is somewhat limited in its ability to get in depth information about model operations.
For example, currently the log is only able to print to stdout, though the information about communications may be useful to have in its raw form,
(ie logging to wandb or tensorboard during training). The logs are also cumulative due it being a global variable. If, for example, we would like
to get information about each specific epoch/iteration, we could not easily do this.

This commit adds the following features
- ability to return communication logs as dictionaries, rather than only printing to stdout
- convenience helper functions for getting information about current logs
- ability to clear existing log operations
- additional documentation for logging operations

Signed-off-by: Alex Kiefer <alexkiefer51@gmail.com>