deepcompile: Call end_backward() from graph
Today, the time to call end_backward() is inferred by counting the
number of gradient blocks that have been calculated. That can be
incorrect for models consisting of multiple modules without leveraging
all their parameters. An example is OpenVLA, which consists of two ViTs
and one LLM and uses the second-to-last layer of both ViTs.

For generality, make backward graphs call end_backward() explicitly at
the end.

Signed-off-by: Junjie Mao <junjie.mao@linux.alibaba.com>