Deduplicate fp32 weights with torch autocast and ZeRO3
When torch autocast is enabled, model weights are already in fp32 and
can be directly updated by the optimizer with fp32 gradients. It is a
waste of accelerator memory to keep another copy, also in fp32, as the
master weight.

Use aliases to the so-called-"fp16" params as the master weights to save
memory. It applies only when no optimizer offloading (either CPU or
NVMe) or swapping mechanisms is enabled.

Signed-off-by: Junjie Mao <junjie.mao@linux.alibaba.com>