deepcompile: do not pass a tuple to release_param()
The release_param() API takes a tensor as its first argument, but a user
of a gathered param may return a tuple when the graph is compiled in
eager mode. That will lead to the following error.

    Profiling error dc::release_param() Expected a value of type 'Tensor' for argument 'a' but instead found type 'tuple'.

Take native_layer_norm as an example, the graph after inserting
allgather and release_param looks like the following.

    %native_layer_norm : [num_users=1] = call_function[target=torch.ops.aten.native_layer_norm.default](args = (%clone, [1024], %wait_allgather_ds_param__primals_8_5, %wait_allgather_ds_param__primals_9_6, 1e-06), kwargs = {})
    %release_ds_param_primals_9_native_layer_norm_6 : [num_users=1] = call_function[target=torch.ops.dc.release_param.default](args = (%native_layer_norm, 140623563718032, 6, 1), kwargs = {})
    %release_ds_param_primals_8_native_layer_norm_5 : [num_users=3] = call_function[target=torch.ops.dc.release_param.default](args = (%release_ds_param_primals_9_native_layer_norm_6, 140623563718032, 5, 1), kwargs = {})
    %getitem : [num_users=1] = call_function[target=operator.getitem](args = (%release_ds_param_primals_8_native_layer_norm_5, 0), kwargs = {})

To fix that exception, check if a param user is consumed by
operator.getitem, and if it is, release the param after the first
getitem call, so that the graph above will now become:

    %native_layer_norm : [num_users=3] = call_function[target=torch.ops.aten.native_layer_norm.default](args = (%clone, [1024], %wait_allgather_ds_param__primals_8_5, %wait_allgather_ds_param__primals_9_6, 1e-06), kwargs = {})
    %getitem : [num_users=1] = call_function[target=operator.getitem](args = (%native_layer_norm, 0), kwargs = {})
    %release_ds_param_primals_9_getitem_6 : [num_users=1] = call_function[target=torch.ops.dc.release_param.default](args = (%getitem, 140032938687600, 6, 1), kwargs = {})
    %release_ds_param_primals_8_getitem_5 : [num_users=1] = call_function[target=torch.ops.dc.release_param.default](args = (%release_ds_param_primals_9_getitem_6, 140032938687600, 5, 1), kwargs = {})

A previous attempt tried to fix that issue by allowing arbitrary type
(i.e. c10::IValue) as the first argument of release_param(), but it was
abandoned because it breaks the integration of deepcompile with inductor
backend. That is because allowing arbitrary type prevents torch
dispatcher from inferring the appropriate dispatch key and somehow
confuses the fake tensor dispatching logic in inductor, making it calls
allgather_param_meta() for allgather_param but release_param() for
release_param.

Signed-off-by: Junjie Mao <junjie.mao@linux.alibaba.com>