add support for tensor learning rate (vs scalar)
This change is intended to help enable support
for using a tensor learning rate value vs a scalar
ones. We found this helpful in cases where the
Optimizer is torch.compiled (in such cases changing
the scalar LR value could cause recompilation degrading
the performance).
The implementation allows the model script to determine the type of
LR value used , by setting the initial value.